% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musica.R
\name{codes}
\alias{code2period}
\alias{codes}
\alias{period2code}
\title{Conversion between period specification and codes}
\usage{
period2code(periods)

code2period(code)
}
\arguments{
\item{periods}{period specification}

\item{code}{period code}
}
\description{
Conversion between period specification and codes
}
\details{
Periods are specified using keywords "day", "month", "year" preceded by an integer and a space and optionally followed by "s" (the specification is further passed to \code{cut.Date}, see \code{\link{cut.Date}} for details). To fit in figures and for simplicity, periods can be also specified by codes, i.e. by \code{D}, \code{M}, \code{Y} (for "day", "month" and "year", respectively) and folowed by integer specifying the number of intervals. The functions \code{period2code} and \code{code2period} provide conversion between the two alternatives.
}
\examples{
period2code(c('1 day', '23 days', '3 month', '2 years'))

code2period(c('D1', 'D23', 'M3', 'Y2'))
}

