% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musica.R
\name{compare}
\alias{compare}
\title{Compare decomposed variables}
\usage{
compare(x, compare_to, fun = mean, wet_int_only = TRUE, wet_int_thr = 0.1,
  exclude_below = 0.9)
}
\arguments{
\item{x}{List of decomposed variables to be compared}

\item{compare_to}{Decomposed variable used as a reference}

\item{fun}{Function used for comparison}

\item{wet_int_only}{(logical) Should only the wet intervals be considered?}

\item{wet_int_thr}{Numeric value specifying the minimum depth to be consider wet}

\item{exclude_below}{Some of the intervals might not be of required length, e.g. D10 interval may have less than 10 days available. The \code{exclude_below} argument controls the minimum fraction of the interval that has to be available in order to be considered in the summary statistics.}
}
\value{
data.table summarizing the differences
}
\description{
Compare decomposed variables
}
\examples{
data(basin_PT)
dobs = decomp(basin_PT[['obs_ctrl']])

}

