% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc.R
\name{correct}
\alias{correct}
\title{Use quantile mapping to correct bias}
\usage{
correct(obs_ctrl, sim_ctrl, sim_scen = copy(sim_ctrl), ...)
}
\arguments{
\item{obs_ctrl}{observed data for the control period}

\item{sim_ctrl}{simulated data for the control period}

\item{sim_scen}{scenario data}

\item{...}{other arguments passed to fitQmap}
}
\value{
data.table with corrected scenario simulation
}
\description{
The function is a simple wrapper around quantile mapping functions provided in the \code{\link{qmap}} package. Quantile mapping is done on each variable included in the input data.table, the mapping is specific for each month.
}
\details{
The correction is calibrated by comparison of \code{obs_ctrl} and \code{sim_ctrl}. Quantile mapping based on empirical quantiles is considered. The correction for wet.day frequency is applied for precipitation.
}
\examples{
data(basin_PT)

# correct future simulation
cor_future = correct(obs_ctrl = basin_PT[['obs_ctrl']], sim_ctrl = basin_PT[['sim_ctrl']], sim_scen =  basin_PT[['sim_scen']])

# get the corrected control simulation
cor_future = correct(obs_ctrl = basin_PT[['obs_ctrl']], sim_ctrl = basin_PT[['sim_ctrl']])
}

