% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musica.R
\name{dcompare}
\alias{dcompare}
\title{Compare distribution function of variables}
\usage{
dcompare(x, compare_to, p = seq(0, 1, 0.01), wet_int_only = TRUE,
  wet_int_thr = 0.1, exclude_below = 0.9)
}
\arguments{
\item{x}{List of decomposed objects}

\item{wet_int_only}{(logical) Should only the wet intervals be considered?}

\item{wet_int_thr}{Numeric value specifying the minimum depth to be consider wet}

\item{exclude_below}{Some of the intervals might not be of required length, e.g. D10 interval may have less than 10 days available. The \code{exclude_below} argument controls the minimum fraction of the interval that has to be available in order to be considered in the summary statistics.}
}
\value{
data.table summarizing the relation with columns:
\describe{
 \item{variable}{factor indicating the variable}
 \item{period}{specification of the averaging length with `D` - day(s), `M` - month(s), `Y` - year(s) and `G1` - the overall mean }
 \item{TS}{averaging length in hours}
 \item{sub_period}{indication of the aggregating scale specified by \code{agg_by} argument}
 \item{comp}{factor indicating the data sets from \code{x} with labels given by \code{names(x)}}
 \item{DIF}{distance between quantiles of \code{x} and \code{compare_to}. Distance is measured as difference for variables included in \code{getOption('additive_variables')}, i.e. temperature (\code{TAS}) by default, and as a ratio for other variables, see \code{\link{dif}}}
}
}
\description{
Compare distribution function of variables
}
\examples{
library(ggplot2)
data(basin_PT)
\dontrun{
dobs = decomp(basin_PT[['obs_ctrl']])
dctrl = decomp(basin_PT[['sim_ctrl']])
d = dcompare(x = list(CTRL = dctrl), compare_to = dobs)
ggplot(d[variable=='TAS' & period!='G1']) +
  geom_line(aes(x = p, y = DIF, col = comp)) +
  facet_grid(sub_period~period, scale = 'free') +
  theme(legend.position = 'top', axis.text.x = element_text(angle = 90, vjust = .5))
}
}

