% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musica.R
\name{decomp}
\alias{decomp}
\title{Decomposition of time-series}
\usage{
decomp(x, year_starts = months(0), scales = c(Y1 = "1 year", M6 = "6 month",
  M3 = "3 months", M1 = "1 months", D5 = "15 days", D1 = "1 day"),
  agg_by = quarter)
}
\arguments{
\item{x}{data.table with columns \code{DTM} (date), \code{variable} and \code{value}.}

\item{year_starts}{A Period object indicating the start of the year}

\item{scales}{The time-scales to be included in output, see Details}

\item{agg_by}{Function for specification of the period (season, month) to be included in output, see Details}
}
\value{
decomposed data.table
}
\description{
Decomposition of time-series
}

