% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr4j.R
\name{gr4j}
\alias{gr4j}
\title{Runoff simulation with the GR4J model}
\usage{
gr4j(data, x = c(350, 0, 40, 0.5), lat = 49.8, return_components = FALSE)
}
\arguments{
\item{data}{data.table with columns DTM (datum), PR (precipitation), TAS (temperature)}

\item{x}{parameters of the GR4J model}

\item{lat}{latitude}
}
\value{
estimated runoff
}
\description{
Runoff simulation with the GR4J model
}

