% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musica.R
\docType{package}
\name{musica-package}
\alias{musica-package}
\title{musica - An R package for multiscale climate model assessment}
\description{
The package provides functions for flexible assessment of climate model bias and changes at multiple time scales. See documentation for \code{\link{decomp}}, \code{\link{compare}} and \code{\link{vcompare}}.
}
\section{Package options}{

Following option(s) are available:
\describe{
 \item{additive_variables}{At several places the package compares values. The character vector \code{additive_values} specifies for which variables difference should be used for comparison instead of ratio. Defaults to \code{additive_values = "TAS"}. See \code{\link{options}} for setting or examining options.}
}
}
\author{
Martin Hanel \email{hanel@fzp.czu.cz}
}
\references{
Hanel, M., Kozin, R. (2016) Bias correction for hydrological modelling, submitted.
}

