% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musica.R
\name{prob}
\alias{prob}
\title{Evaluation of empirical cumulative distribution function}
\usage{
prob(x)
}
\arguments{
\item{x}{vector of values}
}
\value{
value of the empirical distribution function evaluated at \code{x}
}
\description{
Evaluation of empirical cumulative distribution function
}
\examples{
prob(rnorm(10))
}

