% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musica.R
\name{Q}
\alias{Q}
\title{Convenience function for calculation of quantiles}
\usage{
Q(p, ...)
}
\arguments{
\item{p}{Specification of the quantile}

\item{...}{other arguments passed to \code{quantile}}
}
\value{
function calculating the p-th quantile
}
\description{
Convenience function for calculation of quantiles
}

