% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musica.R
\name{sscale2sea}
\alias{sscale2sea}
\title{Provides the letter code for months}
\usage{
sscale2sea(sub_scale, year_starts = months(-1))
}
\arguments{
\item{sub_scale}{Typically the \code{sub_period} variable from decomposed object}

\item{year_starts}{The start of the year}
}
\value{
Vector of three-letter codes for seasons
}
\description{
Provides the letter code for months
}
\details{
Typical workflow is to set \code{year_starts} in the decomp function e.g. to \code{months(-1)} or \code{months(2)}. These both result in climatological seasons (December-January-February - DJF, etc.). The latter in addition results in grouping of warm and cold seasons together at M6 scale. The sub_period field of the decomposed object is with respect to \code{year_starts}, i.e. when \code{year_starts = months(-1)} then \code{sub_period = 1} corresponds to December. To obtain the three-letter codes back, \code{sscale2sea} is used. The function is typically used for plotting.
}
\examples{
sscale2sea(1:12, year_starts = months(-1))
}

