% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr4j.R
\name{t.pet}
\alias{t.pet}
\title{Estimation of potential evapotranspiration based on temperature}
\usage{
\method{t}{pet}(TAS, DTM, lat = 49.8)
}
\arguments{
\item{TAS}{Temperature}

\item{DTM}{datum}

\item{lat}{latitude}
}
\description{
Estimation of potential evapotranspiration based on temperature
}

