% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musica.R
\name{tscale}
\alias{tscale}
\title{Convert temporal scale code to hours}
\usage{
tscale(x, nyears = 30)
}
\arguments{
\item{x}{Vector of the temporal scale codes}

\item{nyears}{Overall number of years - used for conversion of the grand mean}
}
\value{
numerical vector of durations in hours
}
\description{
Interval durations are calculated by the lubridate package.
}
\examples{
tscale('M1')
}

