% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musica.R
\name{vcompare}
\alias{vcompare}
\title{Assess the relations between two decomposed variables}
\usage{
vcompare(x, fun = cor, wet_int_only = TRUE, wet_int_thr = 0.1,
  exclude_below = 0.9)
}
\arguments{
\item{x}{Decomposed object}

\item{fun}{Function to sumarize dependence (like \code{cor}, \code{cov})}

\item{wet_int_only}{(logical) Should only the wet intervals be considered?}

\item{wet_int_thr}{Numeric value specifying the minimum depth to be consider wet}

\item{exclude_below}{Some of the intervals might not be of required length, e.g. D10 interval may have less than 10 days available. The \code{exclude_below} argument controls the minimum fraction of the interval that has to be available in order to be considered in the summary statistics.}
}
\value{
data.table summarizing the relation
}
\description{
Assess the relations between two decomposed variables
}

