% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musica.R
\name{decomp}
\alias{decomp}
\title{Decomposition of time-series}
\usage{
decomp(x, year_starts = months(0), periods = c("1 year", "6 months",
  "3 months", "1 month", "15 days", "1 day"), agg_by = quarter)
}
\arguments{
\item{x}{data.table with columns \code{DTM} (date), \code{variable} and \code{value}. Any number of variables are in principle allowed.}

\item{year_starts}{A Period object indicating the start of the year, e.g. \code{year_starts = -months(1)} results in climatological seasons}

\item{periods}{The periods over which the averages will be calculated, see Details}

\item{agg_by}{Function for specification of the period (season, month) to be additionaly included in output, see Details}
}
\value{
data.table with variables:
\describe{
 \item{variable}{factor indicating the variable}
 \item{DTM}{date}
 \item{period}{specification of the averaging length with `D` - day(s), `M` - month(s), `Y` - year(s) and `G1` - the overall mean }
 \item{value}{value of the variable for given averaging length}
 \item{sub_period}{indication of the aggregating scale specified by \code{agg_by} argument}
 \item{period_pos}{average date of the interval}
 \item{N}{real length of the vectors used for calculating averages}
 \item{TS}{averaging length in hours}
}
}
\description{
Enhance the input data.table with series of averages over the periods specified in the \code{periods} argument.
}
\details{
The original time series in daily time step is decomposed into series of averages ove periods specified in \code{periods} argument using keywords "day", "month", "year" preceded by an integer and a space and optionally followed by "s" (the specification is further passed to \code{cut.Date}, see \code{\link{cut.Date}} for details). The periods must be given in order from longest to shortest, the overall mean is always included (and needs not to be included). Shorter periods are always identified within the closest longer periods, i.e. each shorter period is fully included in exactly one longer period. As a result, the averages may be calculated over shorter periods than specified. This is due to varying length of "month" and "year" periods. The actual length used for averaging is included in the output. To make further assessment of the decomposed objects easier, indicator of period within the year (e.g. quarter or month) as specified by \code{agg_by} argument is included in the output.
}
\examples{
data(basin_PT)
str(basin_PT)
basin_PT[['obs_ctrl']]
dobs = decomp(basin_PT[['obs_ctrl']], scales = c('1 year', '1 month', '1 day'))
}

