% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musica.R
\name{Q}
\alias{Q}
\title{Convenience function for calculation of quantiles}
\usage{
Q(p, ...)
}
\arguments{
\item{p}{Specification of the quantile}

\item{...}{other arguments passed to \code{\link{quantile}}}
}
\value{
function calculating the p-th quantile
}
\description{
The typical use is in \code{\link{compare}} to avoid anonymous functions in specification of its \code{fun} argument.
}
\examples{
q90 = Q(.9)
class(q90)
q90(rnorm(10))
}

